requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cli-helpers==2.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.8
coverage==7.4.1
cryptography==42.0.1
defusedxml==0.8.0rc2
Django==4.2
django-allauth==0.60.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.5
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
execnet==2.0.2
factory-boy==3.3.0
Faker==22.6.0
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
loguru==0.7.2
mccabe==0.7.0
model-bakery==1.17.0
oauthlib==3.2.2
packaging==23.2
pendulum==2.1.2
pgcli==4.0.1
pgspecial==2.1.1
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg==3.1.17
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.4
pytzdata==2020.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
setproctitle==1.3.3
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
